Pretty Logging, Config Saving Implemented
Scrapted Debug script and created Logger Script for pretty logging in the console. Shows time and what process is calling it.
Settings section is style except for the drop down list of profiles to select.
Configuration is based on the UserData folder now and not in the root of the project folder.
Configuration is loaded on start up and is saved whenever there is a keystroke in the username, password, or channel field.
Fixed script issue in package.json that cause "npm start" to execute with the UserData directory being "electron" and not "beam-soundly-interactive".
Added Startup section to the ui.js script for anything that must be fired when the application first loads. (i.e. showing first tab, initializing IPC, and loading config)
Changed all the const to vars, as we don't need to be in strict mode for what we are doing. (That might change later)