Modify/Save/Delete profiles
Profile sounds(title & url) can now be edited and are saved based on the profile you have selected in the settings menu.
The Logger has been updated to use utils to do a better job at dumping objects into the console for debug purposes.
Config functionality in the background has all features directly related to the UI plus the additional ability to create a new profile which is not yet implimented on the front end.
Fixed tag breaks in the index.html. Uncessary breaks from an atom plugin were just wasting space.